+++ src/query.rs
@@ -612,6 +612,9 @@ impl<'w, Q: Query> QueryBorrow<'w, Q> {
             return;
         }
         for x in self.archetypes {
+            if x.is_empty() {
+                continue;
+            }
             // TODO: Release prior borrows on failure?
             if let Some(state) = Q::Fetch::prepare(x) {
                 Q::Fetch::borrow(x, state);
@@ -689,6 +692,9 @@ impl<'w, Q: Query> Drop for QueryBorrow<'w, Q> {
     fn drop(&mut self) {
         if self.borrowed {
             for x in self.archetypes {
+                if x.is_empty() {
+                    continue;
+                }
                 if let Some(state) = Q::Fetch::prepare(x) {
                     Q::Fetch::release(x, state);
                 }
@@ -1147,6 +1153,9 @@ impl<'q, Q: Query> PreparedQueryBorrow<'q, Q> {
         fetch: &'q mut [Option<Q::Fetch>],
     ) -> Self {
         for (idx, state) in state {
+            if archetypes[*idx].is_empty() {
+                continue;
+            }
             Q::Fetch::borrow(&archetypes[*idx], *state);
         }
 
@@ -1179,6 +1188,9 @@ impl<'q, Q: Query> PreparedQueryBorrow<'q, Q> {
 impl<Q: Query> Drop for PreparedQueryBorrow<'_, Q> {
     fn drop(&mut self) {
         for (idx, state) in self.state {
+            if self.archetypes[*idx].is_empty() {
+                continue;
+            }
             Q::Fetch::release(&self.archetypes[*idx], *state);
         }
     }